upd jar file and fix parser options
