+++ src/main/java/allure/to/docx/Runner.java
@@ -1,7 +1,10 @@
 package allure.to.docx;
 
-import allure.to.docx.core.DocxHelper;
-import allure.to.docx.core.ResultHelper;
+import allure.to.docx.core.DocxCreator;
+import allure.to.docx.core.ResultParser;
+import allure.to.docx.options.Options;
+import allure.to.docx.util.LocaleUtil;
+import com.google.devtools.common.options.OptionsParser;
 import lombok.extern.java.Log;
 
 import java.io.File;
@@ -10,12 +13,14 @@ import java.io.File;
 public class Runner {
 
     public static void main(final String[] args) {
+        OptionsParser parser = OptionsParser.newOptionsParser(Options.class);
+        parser.parseAndExitUponError(args);
+        var options = parser.getOptions(Options.class);
 
-        var from = System.getProperty("from", "./allure-results");
-        var out = System.getProperty("out", "out.docx");
+        LocaleUtil.initialize(options.language, options.country);
 
-        var results = ResultHelper.getTestResults(new File(from).toPath());
-        DocxHelper.saveDocument(DocxHelper.createDocxFile(results), out);
+        var results = ResultParser.getTestResults(new File(options.from).toPath());
+        DocxCreator.saveDocument(DocxCreator.createDocxFile(results), options.output);
     }
 
 }
