--- src/main/resources/l10n.properties
@@ -1,12 +1,0 @@
-test.report=Отчет по тестированию
-test.period=Результаты тестирования за 
-test.cases=Тест кейсы
-test.case=Тест кейс
-tests.total=Всего: 
-tests.successful=Успешных: 
-tests.failed=Проваленных: 
-tests.broken=Сломанных: 
-tests.skipped=Пропущенных: 
-tests.unknown=Неизвестных: 
-test.status=Статус
-pie.chart.title=Из %d тестов успешных %d
